+++ src/pages/index.tsx
@@ -17,7 +17,7 @@ import {
   VStack,
   useToast,
 } from '@chakra-ui/react';
-import { LinkIcon, SearchIcon } from '@chakra-ui/icons';
+import { LinkIcon } from '@chakra-ui/icons';
 import Logo from '../components/Logo';
 import PageLayout from '../layout/Page';
 import DevName from '../components/Search/Dev/DevName';
@@ -38,6 +38,7 @@ function App() {
     (e) => {
       if (e <= 8000) {
         setDeveloperId(e);
+        router.replace({ query: { id: e } });
       }
     },
     [router],
@@ -46,8 +47,7 @@ function App() {
   function getSearchID() {
     if (process.browser) {
       const search = window.location.search;
-      const val = new URLSearchParams(search).get('id') || 1;
-      return val;
+      return new URLSearchParams(search).get('id') || 1;
     }
     return 1;
   }
@@ -68,17 +68,82 @@ function App() {
               name="hero-field"
               bg="white"
             />
-            <Button
-              as="a"
-              href={`/id/${developerId}`}
-              leftIcon={<SearchIcon />}
-            >
-              Search
-            </Button>
           </VStack>
+          {typeof window !== 'undefined' ? (
+            <NftProvider fetcher={['ethers', ethersConfig]}>
+              <Nft developerId={developerId.toString()} />
+            </NftProvider>
+          ) : (
+            <Text>{t('loading')}</Text>
+          )}
         </VStack>
       </chakra.main>
     </PageLayout>
+  );
+}
+
+function Nft({ developerId }: { developerId: string }) {
+  const { t } = useTranslation();
+  const toast = useToast();
+
+  const copyLinkToNFT = useCallback(() => {
+    navigator.clipboard.writeText(`${SITE_URL}/?id=${developerId}`);
+    toast({
+      title: t('linkCopied'),
+      isClosable: true,
+    });
+  }, [toast, t, developerId]);
+
+  const { loading, error, nft } = useNft(DEVELOPER_DAO_CONTRACT, developerId);
+
+  const [nftImage, nftAltText] = useNftImageContent(nft?.image);
+
+  if (loading) return <Text>{t('loading')}</Text>;
+
+  if (!developerId) return <Text>{t('enterDeveloperId')}</Text>;
+
+  if (parseInt(developerId) > 8000) return <Text>{t('invalidToken')}</Text>;
+
+  if (error || !nft) return <Text>{t('error')}.</Text>;
+
+  return (
+    <VStack w="full" spacing={5}>
+      <chakra.img
+        alt={nftAltText!}
+        src={nftImage!}
+        border={4}
+        borderStyle="solid"
+        borderColor="gray.200"
+        w="full"
+        objectFit="cover"
+        objectPosition="center"
+        rounded="md"
+      />
+      <VStack>
+        <DevName nft={nft} developerId={developerId} />
+        {nft.owner ? (
+          <Button
+            as="a"
+            href={`${ETHER_SCAN_LINK_PREFIX}/${nft.owner}`}
+            target="_blank"
+            rel="noreferrer"
+            title={t('viewOwnerEtherscan')}
+            fontSize={{ base: 'xs', sm: 'md' }}
+          >
+            {t('owner')}:&nbsp;
+            <chakra.span maxW="xs">{nft.owner.slice(0, 30)}</chakra.span>...
+            {nft.owner.slice(-4)}
+          </Button>
+        ) : (
+          <Button isDisabled>
+            {t('owner')}:&nbsp;{t('unclaimed')}
+          </Button>
+        )}
+        <Button onClick={copyLinkToNFT} leftIcon={<LinkIcon />}>
+          {t('copyLinkToNFT')}
+        </Button>
+      </VStack>
+    </VStack>
   );
 }
